requirements.txt
pyrogram>=2.0.106
TgCrypto>=1.2.5
python-dotenv>=1.0.0
aiogram>=3.4.1 